# Improve memory throughput of atom by atom