--- a/timemachine/cpp/src/bd_exchange_move.cu
+++ b/timemachine/cpp/src/bd_exchange_move.cu
@@ -313,7 +313,8 @@ void BDExchangeMove<RealType>::compute_incremental_log_weights_device(
     const RealType *d_translations, // [batch_size_, 3]
     cudaStream_t stream) {
     const int tpb = DEFAULT_THREADS_PER_BLOCK;
-    dim3 atom_by_atom_grid(ceil_divide(N, tpb), mol_size_ * batch_size_, 1);
+    // Scale down the y block by 16 to have each block do more work. Improves memory throughput
+    dim3 atom_by_atom_grid(ceil_divide(N, tpb), ceil_divide(mol_size_ * batch_size_, 16), 1);
 
     k_setup_proposals<<<ceil_divide(batch_size_, tpb), tpb, 0, stream>>>(
         num_proposals_per_move_,